% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{add_credible_intervals}
\alias{add_credible_intervals}
\title{add_credible_intervals}
\usage{
add_credible_intervals(x, ...)
}
\arguments{
\item{x}{Input object.}

\item{...}{Extra arguments.}
}
\description{
Generic for add_credible_intervals().
}

