% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{calculate_average_performance}
\alias{calculate_average_performance}
\title{calculate_average_performance}
\usage{
calculate_average_performance(happy_metrics, metrics = c("METRIC.Recall",
  "METRIC.Precision", "METRIC.Frac_Assessed", "METRIC.F1_Score"),
  filter = "PASS")
}
\arguments{
\item{happy_metrics}{A `data.frame` with happy metrics obtained with `load_happy_metrics()`.}

\item{metrics}{A `vector` with the metrics to consider.}

\item{filter}{Variant filter. Default = PASS.}
}
\value{
A summarised `data.frame`.
}
\description{
Calculate summary stats for the specified metrics.
}
\examples{
\dontrun{calculate_average_performance(happy_metrics)}
}

