% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{compare_hdi}
\alias{compare_hdi}
\title{Compare subset-specific Beta distributions across two groups}
\usage{
compare_hdi(happy_hdi, sample_size = 1e+05, significance = 0.05)
}
\arguments{
\item{happy_hdi}{A \code{happy_hdi} object obtained with \code{estimate_hdi()}.}

\item{sample_size}{Number of observations to draw from the Beta posterior
to estimate HDIs for the difference in Beta distributions. Default: 1e5.}

\item{significance}{Significance levels for HDIs of the difference. Default: 0.05 (= 95\% HDIs).}
}
\value{
A \code{data.frame} with estimated differences across groups, HDIs for the difference
and significance labels.
}
\description{
For each subset, calculate the difference in Beta distributions across two groups, and label it as significant if
we can confidently claim that the HDI of the difference is other than zero. Do not attempt tests
in subsets for which we do not have a good estimate of the success rate to start with (HDI range = 1).
}
\examples{

\dontrun{
hdi <- estimate_hdi(df, successes_col = 'TRUTH.TP', totals_col = 'TRUTH.TOTAL', 
                   group_cols = c('Group.Id', 'Subset', 'Type', 'Subtype'))
hdi_diff <- compare_hdi(happy_hdi = hdi)                 
}

}
