% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_happy_summary}
\alias{demo_happy_summary}
\title{demo_happy_summary}
\format{A `happy_compare` object (i.e. a `list`) with the following elements:
\describe{
  \item{config}{Original config (a `data.table`).}
  \item{happy_summary}{Happy summary metrics (a `happy_summary` object, i.e. a `list` of `data.tables`.)}
  \item{happy_extended}{Happy extended metrics (a `happy_extended` object, i.e. a `list` of `data.tables`.)}
}}
\usage{
demo_happy_summary
}
\description{
Demo data for the happy summary class
}
\keyword{datasets}

