% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happyCompare_list.R
\name{extract}
\alias{extract}
\title{Extract tables from a happyCompare_list object}
\usage{
extract(happyCompare_list, ...)
}
\arguments{
\item{happyCompare_list}{A \code{happyCompare_list} object.}

\item{table}{Table of data to extract from each \code{happy_result} object.
Must be one of: \code{summary}, \code{extended}, \code{pr_curve.all},
\code{pr_curve.SNP},  \code{pr_curve.SNP_PASS}, \code{pr_curve.SNP_SEL},
\code{pr_curve.INDEL},  \code{pr_curve.INDEL_PASS},
\code{pr_curve.INDEL_SEL}.}
}
\value{
A \code{data.frame} with combined information from the selected
  \code{happy_result} and its \code{samplesheet}.
}
\description{
Extract tables from a `happyCompare_list object` and combine into a single
\code{data.frame}. Appends samplesheet metadata to the original happy
results.
}
\examples{

\dontrun{
summary = extract(happyCompare_list, table = "summary")
roc = extract(happyCompare_list, table = "pr_curve.all")
}
}

