% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{extract_metrics}
\alias{extract_metrics}
\title{Extract tables from a happy_compare object}
\usage{
extract_metrics(happy_compare, table = c("summary", "extended", "pr.all",
  "pr.snp.all", "pr.snp.pass", "pr.snp.sel", "pr.indel.all", "pr.indel.pass",
  "pr.indel.sel"))
}
\arguments{
\item{happy_compare}{A \code{happy_compare} object.}

\item{table}{Table of data to extract from each \code{happy_result} object.
Must be one of: \code{summary}, \code{extended}, \code{pr.all},
\code{pr.snp.all}, \code{pr.snp.pass}, \code{pr.snp.sel},
\code{pr.indel.all}, \code{pr.indel.pass}, \code{pr.indel.sel}.}
}
\value{
A \code{data.frame} with combined information from the selected
\code{happy_result} and its \code{samplesheet}.
}
\description{
Extract tables from a \code{happy_compare} object and combine into a single
\code{data.frame}. Appends samplesheet metadata to the original happy
results.
}
\examples{

\dontrun{
summary = extract_metrics(happy_compare, table = 'summary')
roc = extract_metrics(happy_compare, table = 'pr.snp.pass')
}

}
