% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{happy_result_list}
\alias{happy_result_list}
\title{happy_result_list}
\description{
Demo `happy_result_list` including 2 `happy_result` objects.
Generated from hap.py demo data (see `helper_demo_data.R`.
}

