% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{hc_plot_hdi}
\alias{hc_plot_hdi}
\title{Plot HDI}
\usage{
hc_plot_hdi(happy_hdi, title = NULL)
}
\arguments{
\item{happy_hdi}{A \code{data.frame} obtained with \code{estimate_hdi()}.}

\item{title}{Title. Default: \code{NULL}.}
}
\description{
Plot distribution of success rates and HDI estimates for
a given subset from a \code{data.frame}.
}
\examples{

\dontrun{
hdi = estimate_hdi(happy_extended, successes_col = 'TRUTH.TP', totals_col = 'TRUTH.TOTAL', 
                   group_cols = c('Group.Id', 'Subset', 'Type', 'Subtype'))
hc_plot_hdi(happy_hdi = hdi \%>\% filter(Subset == '*'))  
}
}
