% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{hc_summarise_metrics}
\alias{hc_summarise_metrics}
\title{Summarise metrics}
\usage{
hc_summarise_metrics(df, group_cols = c("Group.Id", "Type", "Filter"),
  metric_cols = c("METRIC.F1_Score", "METRIC.Recall", "METRIC.Precision",
  "METRIC.Frac_NA"), digits = 4)
}
\arguments{
\item{df}{A \code{data.frame} object.}

\item{group_cols}{Vector of columns to group counts by. Observations
within the same group will be treated as replicates.}

\item{metric_cols}{Vector of metric columns for which to calculate summary stats.}

\item{digits}{Number of decimal places. Default: \code{4}.}
}
\value{
A formatted \code{data.frame} object.
}
\description{
Calculate mean and sd for specified metrics and return formatted results.
}
\examples{

\dontrun{
df <- extract_metrics(happy_compare, table = "summary")
hc_summarise_metrics(df)
}

}
