% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{hc_tabulate_summary}
\alias{hc_tabulate_summary}
\title{Tabulate hap.py summary results}
\usage{
hc_tabulate_summary(happy_summary, type = c("SNP", "INDEL"),
  filter = c("PASS", "ALL"), cols, colnames = cols, digits = 4,
  aggregate = FALSE)
}
\arguments{
\item{happy_summary}{A \code{happy_summary} object.}

\item{type}{Variant type. One of: \code{SNP}, \code{INDEL}.}

\item{filter}{Variant filter. One of: \code{PASS}, \code{ALL}.}

\item{cols}{Columns to keep.}

\item{colnames}{Column names to use in the output table. Defaults to
\code{cols}.}

\item{digits}{Number of digits to display for performance
metrics. Default: \code{4}.}

\item{aggregate}{Summarise performance across groups. Default: \code{FALSE}.}
}
\value{
A \code{data.frame} object with the selected hap.py performance
metrics.
}
\description{
Display hap.py summary results into tabular format using \code{knitr}.
}
\examples{

\dontrun{
cols = c('Group.Id', 'Replicate.Id', 'METRIC.Recall', 'METRIC.Precision')
colnames = c('Group.Id', 'Sample', 'Recall', 'Precision')
t = hc_tabulate_summary(happy_summary, cols = cols, colnames = colnames, 
             filter = 'PASS', vartype = 'SNP', aggregate = FALSE)
}
}
