% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{HCM.ConfidenceIntervals}
\alias{HCM.ConfidenceIntervals}
\title{HCM.ConfidenceIntervals}
\usage{
HCM.ConfidenceIntervals(subsetted_happy_metrics, variant_type,
  filter = "PASS")
}
\arguments{
\item{subsetted_happy_metrics}{A `data.frame` with happy metrics obtained with `load_happy_metrics()`, 
subsetted to contain a single subset id.}

\item{variant_type}{Variant type. Default = NULL.}

\item{filter}{Variant filter. Default = PASS.}
}
\value{
A `list` of results.
}
\description{
Plot average and confident intervals for performance metrics in a given subset.
}
\examples{
\dontrun{HCM.ConfidenceIntervals(subsetted_happy_metrics = subsetted_happy_metrics, variant_type = 'SNP')}
}

