% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmp.R
\name{HCM.StratifiedPerformance}
\alias{HCM.StratifiedPerformance}
\title{HCM.StratifiedPerformance}
\usage{
HCM.StratifiedPerformance(happy_metrics, title = "Stratified performance")
}
\arguments{
\item{happy_metrics}{A `data.frame` with happy metrics obtained with `load_happy_metrics()`.}

\item{title}{Plot title. Default: 'Stratified performance'.}
}
\value{
A `list` of results.
}
\description{
Plot flowcell layout for builds in each group.
}
\examples{
\dontrun{PCM.TwoGroupFlowcellLayout(groupA, groupB)}
}

