% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCM.TopDifferingSubsets.R
\name{HCM.TopDifferingSubsets}
\alias{HCM.TopDifferingSubsets}
\title{HCM.TopDifferingSubsets}
\usage{
HCM.TopDifferingSubsets(happy_metrics, filter = "PASS", variant_type = NULL)
}
\arguments{
\item{happy_metrics}{A `data.frame` with happy metrics obtained with `load_happy_metrics()`.}

\item{filter}{Variant filter. Default = PASS.}

\item{variant_type}{Variant type. Default = NULL.}
}
\value{
A `list` of results.
}
\description{
Evaluate differences in performance metrics for subsets in `happy_metrics` and flag top differing ones.
}
\examples{
\dontrun{HCM.TopDifferingSubsets(happy_metrics)}
}

