% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metrics.R
\name{is_build_metrics}
\alias{is_build_metrics}
\title{is_build_metrics}
\usage{
is_build_metrics(x)
}
\arguments{
\item{x}{An object to inspect.}
}
\description{
Check if the class of the provided object matches the expected one.
}

