% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplocompare.R
\name{is_haplocompare}
\alias{is_haplocompare}
\title{is_haplocompare}
\usage{
is_haplocompare(x)
}
\arguments{
\item{x}{An object to inspect.}
}
\description{
Check if the class of the provided object matches the expected one.
}

