% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_extended_ci.R
\name{is_happy_extended_ci}
\alias{is_happy_extended_ci}
\title{is_happy_extended_ci}
\usage{
is_happy_extended_ci(x)
}
\arguments{
\item{x}{An object to inspect.}
}
\description{
Check if the class of the provided object matches the expected one.
}

