% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_roc.R
\name{is_happy_roc}
\alias{is_happy_roc}
\title{is_happy_roc}
\usage{
is_happy_roc(x)
}
\arguments{
\item{x}{An object to inspect.}
}
\description{
Check if the class of the provided object matches the expected one.
}

