% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_summary.R
\name{is_happy_summary}
\alias{is_happy_summary}
\title{is_happy_summary}
\usage{
is_happy_summary(x)
}
\arguments{
\item{x}{An object to inspect.}
}
\description{
Check if the class of the provided object matches the expected one.
}

