% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sompy_stats.R
\name{is_sompy_stats}
\alias{is_sompy_stats}
\title{is_sompy_stats}
\usage{
is_sompy_stats(x)
}
\arguments{
\item{x}{An object to inspect.}
}
\description{
Check if the class of the provided object matches the expected one.
}

