% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{load_data}
\alias{load_data}
\title{load_data}
\usage{
load_data(config_path)
}
\arguments{
\item{config_path}{Path to configuration file. Required columns:
group_id, sample_id, happy_summary, happy_extended.}
}
\value{
A `haplocompare` object.
}
\description{
Load csv inputs.
}
\examples{
\dontrun{load_data(config_path = config_path)}
}

