% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{load_happy_metrics}
\alias{load_happy_metrics}
\title{load_happy_metrics}
\usage{
load_happy_metrics(config_path, L1_subsets = NULL)
}
\arguments{
\item{config_path}{Path to configuration file. Required columns:
group (group name), id (results id),
happy_csv (path to happy *.summary.csv or *.extended.csv files).}

\item{L1_subsets}{A `vector` containing subsets to be classified as level 1.}
}
\value{
A combined `data.frame`.
}
\description{
Load happy csv output into a `data.frame`.
}
\examples{
\dontrun{load_happy_metrics(config_path = config_path)}
}

