% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplocompare.R
\name{merge_nested.haplocompare}
\alias{merge_nested.haplocompare}
\title{merge_nested.haplocompare}
\usage{
\method{merge_nested}{haplocompare}(obj, x, y, select, ...)
}
\arguments{
\item{obj}{A `haplocompare` object.}

\item{x}{First nested object to merge. One of: `names(haplocompare_obj)`.}

\item{y}{Second nested object to merge. One of: `names(haplocompare_obj)`.}

\item{select}{Columns to keep after merge.}

\item{...}{Extra arguments.}
}
\value{
A merged `data.table`.
}
\description{
Merge contents of a haplocompare object, using the combination of Group.Id,
Sample.Id and Replicate.Id as unique identifiers.
}

