% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_results.R
\name{plot.happy_roc}
\alias{plot.happy_roc}
\title{Plot a ROC curve for precision vs recall}
\usage{
\method{plot}{happy_roc}(happy_roc, filter = "PASS", type, subtype = "*",
  subset = "*", legend_position = "bottom", ...)
}
\arguments{
\item{happy_roc}{A \code{happy_roc} object.}

\item{filter}{Variant filter. Default: \code{PASS}.}

\item{type}{Variant type. One of: \code{SNP}, \code{INDEL}.}

\item{subtype}{Variant subtype. Default: \code{*}.}

\item{subset}{Variant subset. Default: \code{*}.}

\item{legend_position}{A character string to pass to \code{ggplot} to specify
the legend position. Default: \code{bottom}.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot precision vs recall ROC curves from the selected filters.
}
\examples{

\dontrun{
roc = extract(happyCompare_list, table = "pr_curve.all")
plot(roc, type = "SNP", filter = "ALL")  
}
}

