% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_summary.R
\name{plot.happy_summary}
\alias{plot.happy_summary}
\title{plot.happy_summary}
\usage{
\method{plot}{happy_summary}(x, type, filter = "PASS", xlim.low = NA,
  xlim.high = 1, ylim.low = NA, ylim.high = 1, point.size = 3,
  font.size = 12, ...)
}
\arguments{
\item{x}{A `happy_summary` object.}

\item{type}{Variant type. One of: SNP, INDEL.}

\item{filter}{Variant filter. One of: PASS (default), ALL.}

\item{xlim.low}{Lower bound for x axis. Default: NA.}

\item{xlim.high}{Upper bound for x axis. Default: 1.}

\item{ylim.low}{Lower bound for y axis. Default: NA.}

\item{ylim.high}{Upper bound for y axis. Default: 1.}

\item{point.size}{Point size in scatterplot. Default: 3.}

\item{font.size}{Font size. Default: 12.}

\item{...}{Extra arguments.}
}
\description{
Plot happy summary results.
}

