% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sompy_stats.R
\name{plot.sompy_stats}
\alias{plot.sompy_stats}
\title{plot.sompy_stats}
\usage{
\method{plot}{sompy_stats}(x, type, filter = "PASS", xlim.low = 0,
  xlim.high = 1, ylim.low = 0, ylim.high = NA, point.size = 3,
  font.size = 12, ...)
}
\arguments{
\item{x}{A `sompy_stats` object.}

\item{type}{Variant type. One of: SNVs, indels.}

\item{filter}{Variant filter, only used in the plot title.}

\item{xlim.low}{Lower bound for x axis. Default: 0.}

\item{xlim.high}{Upper bound for x axis. Default: 1.}

\item{ylim.low}{Lower bound for y axis. Default: 0.}

\item{ylim.high}{Upper bound for y axis. Default: NA.}

\item{point.size}{Point size in scatterplot. Default: 3.}

\item{font.size}{Font size. Default: 12.}

\item{...}{Extra arguments.}
}
\description{
Plot sompy stats results.
}

