% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_extended_ci.R
\name{plot_subset.happy_extended_ci}
\alias{plot_subset.happy_extended_ci}
\title{plot_subset}
\usage{
\method{plot_subset}{happy_extended_ci}(obj, type, filter = "PASS", subset,
  metric, xlim.low = 0, xlim.high = 1, ...)
}
\arguments{
\item{obj}{A `happy_extended_ci` object.}

\item{type}{Variant type. One of: SNP, INDEL.}

\item{filter}{Variant filter. One of: PASS (default), ALL.}

\item{subset}{A subset id.}

\item{xlim.low}{Lower bound for x axis. Default: NA.}

\item{xlim.high}{Upper bound for x axis. Default: 1.}
}
\description{
Plot performance metrics for the specified subset
}

