% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metrics.R
\name{print.build_metrics}
\alias{print.build_metrics}
\title{print.build_metrics}
\usage{
\method{print}{build_metrics}(x, ...)
}
\arguments{
\item{x}{A `build_metrics` object.}

\item{...}{Extra arguments.}
}
\description{
Print a build_metrics object.
}

