% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplocompare.R
\name{print.haplocompare}
\alias{print.haplocompare}
\title{print.haplocompare}
\usage{
\method{print}{haplocompare}(x, ...)
}
\arguments{
\item{x}{A `haplocompare` object.}

\item{...}{Extra arguments.}
}
\description{
Print a haplocompare object.
}

