% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_extended.R
\name{print.happy_extended}
\alias{print.happy_extended}
\title{print.happy_extended}
\usage{
\method{print}{happy_extended}(x, ...)
}
\arguments{
\item{x}{A `happy_extended` object.}

\item{...}{Extra arguments.}
}
\description{
Print a happy_extended object.
}

