% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_roc.R
\name{print.happy_roc}
\alias{print.happy_roc}
\title{print.happy_roc}
\usage{
\method{print}{happy_roc}(x, ...)
}
\arguments{
\item{x}{A `happy_roc` object.}

\item{...}{Extra arguments.}
}
\description{
Print a happy_roc object.
}

