% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_summary.R
\name{print.happy_summary}
\alias{print.happy_summary}
\title{print.happy_summary}
\usage{
\method{print}{happy_summary}(x, ...)
}
\arguments{
\item{x}{A `happy_summary` object.}

\item{...}{Extra arguments.}
}
\description{
Print a happy_summary object.
}

