% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sompy_stats.R
\name{print.sompy_stats}
\alias{print.sompy_stats}
\title{print.sompy_stats}
\usage{
\method{print}{sompy_stats}(x, ...)
}
\arguments{
\item{x}{A `sompy_stats` object.}

\item{...}{Extra arguments.}
}
\description{
Print a sompy_stats object.
}

