% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happyCompare_list.R
\name{read_samplesheet}
\alias{read_samplesheet}
\title{Load data from happyCompare samplesheets}
\usage{
read_samplesheet(samplesheet_path, lazy = TRUE)
}
\arguments{
\item{samplesheet_path}{Path to a happyCompare samplesheet. Required fields:
\code{Group.Id}, \code{Sample.Id}, \code{Replicate.Id},
\code{happy_prefix}.}

\item{lazy}{Do not load larger hap.py results until needed. Default: \code{TRUE}.}
}
\value{
A \code{happyCompare_list} object, with the following fields: 
  \itemize{ \item{\code{samplesheet}: the original samplesheet, stored as a
  \code{data.frame}}. \item{\code{happy_results}: a \code{happy_results_list}
  object that contains individual \code{happy_result} objects as defined in
  \code{happyR}}. }
}
\description{
Load hap.py results into a \code{happyCompare_list} object as specified in
the \code{happyCompare} samplesheet provided. Depends on \code{happyR}.
}
\examples{

\dontrun{
happyCompare_list = read_samplesheet(samplesheet_path = "happyCompare_samplesheet.csv")
}

}

