% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metrics.R
\name{rename.build_metrics}
\alias{rename.build_metrics}
\title{rename}
\usage{
rename.build_metrics(x, metrics_map, ...)
}
\arguments{
\item{x}{A `build_metrics` object.}

\item{metrics_map}{A `data.frame` or `data.table`describing the mapping
between metric names. Required columns: old_name, new_name.}

\item{...}{Extra arguments.}
}
\description{
Rename metric names in a `build_metrics` object.
}

