% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rename_metrics}
\alias{rename_metrics}
\title{rename_metrics}
\usage{
rename_metrics(x, ...)
}
\arguments{
\item{x}{Input object.}

\item{...}{Extra arguments.}
}
\description{
Generic for rename_metrics().
}

