% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sompy_stats.R
\name{summary.sompy_stats}
\alias{summary.sompy_stats}
\title{summary.sompy_stats}
\usage{
\method{summary}{sompy_stats}(object, type, filter = "PASS", digits = 4,
  kable_format = "markdown", ...)
}
\arguments{
\item{object}{A `sompy_stats` object.}

\item{type}{Variant type. One of: SNVs, indels.}

\item{filter}{Variant filter, only used in the table caption.}

\item{digits}{Number of significant digits in summary statistics. Default: 4.}

\item{kable_format}{a character string to pass to `knitr::kable()`. Default: markdown.}

\item{...}{Extra arguments.}
}
\description{
Summarise happy summary results into tabular format.
}

