% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_results.R
\name{tabulate}
\alias{tabulate}
\title{Tabulate hap.py summary results}
\usage{
tabulate(happy_summary, ...)
}
\arguments{
\item{happy_summary}{A \code{happy_summary} object.}

\item{type}{Variant type. One of: \code{SNP}, \code{INDEL}.}

\item{filter}{Variant filter. One of: \code{PASS}, \code{ALL}.}

\item{cols}{Columns to keep.}

\item{colnames}{Column names to use in the output table. Defaults to
\code{cols}.}

\item{significant_digits}{Number of significant digits in performance
metrics. Default: \code{4}.}

\item{aggregate}{Summarise performance across groups. Default: \code{FALSE}.}
}
\value{
A \code{data.frame} object with the selected hap.py performance
  metrics.
}
\description{
Display hap.py summary results into tabular format using \code{knitr}.
}
\examples{

\dontrun{
cols = c("Group.Id", "Replicate.Id", "METRIC.Recall", "METRIC.Precision")
colnames = c("Group.Id", "Sample", "Recall", "Precision")
t = tabulate(happy_summary = happy_summary, cols = cols, colnames = colnames, 
             filter = "PASS", vartype = "SNP", aggregate = FALSE)
}
}

