% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_metrics.R
\name{tidy.build_metrics}
\alias{tidy.build_metrics}
\title{tidy}
\usage{
\method{tidy}{build_metrics}(x, metrics, ...)
}
\arguments{
\item{x}{A `build_metrics` object.}

\item{metrics}{A vector with the metric names to extract.}

\item{...}{Extra arguments.}
}
\description{
Tidy a `build_metrics` object by converting into a single `data.table`.
}

