% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_extended.R
\name{tidy.happy_extended}
\alias{tidy.happy_extended}
\title{tidy}
\usage{
\method{tidy}{happy_extended}(obj)
}
\arguments{
\item{obj}{A `happy_extended` object.}
}
\description{
Tidy a `happy_extended` object by converting into a single `data.table`.
}

