% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_roc.R
\name{tidy.happy_roc}
\alias{tidy.happy_roc}
\title{tidy}
\usage{
\method{tidy}{happy_roc}(x, ...)
}
\arguments{
\item{x}{A `happy_roc` object.}

\item{...}{Extra arguments.}
}
\description{
Tidy a `happy_roc` object by converting into a single `data.table`.
}

