% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_summary.R
\name{tidy.happy_summary}
\alias{tidy.happy_summary}
\title{tidy}
\usage{
\method{tidy}{happy_summary}(obj, ...)
}
\arguments{
\item{obj}{A `happy_summary` object.}
}
\description{
Tidy a `happy_summary` object by converting into a single `data.table`.
}

