% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sompy_stats.R
\name{tidy.sompy_stats}
\alias{tidy.sompy_stats}
\title{tidy}
\usage{
\method{tidy}{sompy_stats}(x, ...)
}
\arguments{
\item{x}{A `sompy_stats` object.}

\item{...}{Extra arguments.}
}
\description{
Tidy a `sompy_stats` object by converting into a single `data.table`.
}

