% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_extended.R
\name{add_credible_intervals.happy_extended}
\alias{add_credible_intervals.happy_extended}
\title{add_credible_intervals}
\usage{
\method{add_credible_intervals}{happy_extended}(obj, metric, sig = 0.05,
  samplesize = 1e+06)
}
\arguments{
\item{obj}{A `happy_extended` object.}
}
\value{
A `data.table` object.
}
\description{
Add credible intervals to the observed performance estimates using a 
beta-binomial model. The uncertainty measurement takes into account both the total number of events in each subset
(e.g. `TRUTH.TOTAL`) and the variability across replicates.
}

