% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{estimate_hdi}
\alias{estimate_hdi}
\title{Estimate HDIs using Beta-Binomial inference}
\usage{
estimate_hdi(happy_extended, successes_col, totals_col, group_cols,
  significance = 0.05, sample_size = 1e+05, max_alpha1 = 1000)
}
\arguments{
\item{happy_extended}{A \code{happy_extended} object.}

\item{successes_col}{Name of the column that contains success counts.}

\item{totals_col}{Name of the column that contains total counts.}

\item{group_cols}{Vector of columns to group counts by. Observations within the same group will be treated as replicates.}

\item{significance}{Significance for HDI estimation. Default: 0.05 (= 95% HDIs).}

\item{sample_size}{Number of observations to draw from the Beta posterior to estimate HDIs. Default: 1e5.}

\item{max_alpha1}{Upper bound for alpha hyperparameter in the aggregate posterior Beta.}
}
\value{
A \code{data.frame} with combined information from the selected
  \code{happy_result} and its \code{samplesheet}.
}
\description{
Estimate HDIs and success rates from hap.py counts using Beta-Binomial inference
}
\examples{

\dontrun{
hdi = estimate_hdi(happy_extended, successes_col = "TRUTH.TP", totals_col = "TRUTH.TOTAL", 
                   group_cols = c("Group.Id", "Subset", "Type", "Subtype"))
}

}
