% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{extract}
\alias{extract}
\title{Extract tables from a happy_compare object}
\usage{
extract(happy_compare, table)
}
\arguments{
\item{happy_compare}{A \code{happy_compare} object.}

\item{table}{Table of data to extract from each \code{happy_result} object.
Must be one of: \code{summary}, \code{extended}, \code{pr_curve.all},
\code{pr_curve.SNP},  \code{pr_curve.SNP_PASS}, \code{pr_curve.SNP_SEL},
\code{pr_curve.INDEL},  \code{pr_curve.INDEL_PASS},
\code{pr_curve.INDEL_SEL}.}
}
\value{
A \code{data.frame} with combined information from the selected
\code{happy_result} and its \code{samplesheet}.
}
\description{
Extract tables from a \code{happy_compare} object and combine into a single
\code{data.frame}. Appends samplesheet metadata to the original happy
results.
}
\examples{

\dontrun{
summary = extract(happy_compare, table = 'summary')
roc = extract(happy_compare, table = 'pr_curve.all')
}

}
