% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{hc_plot_hdi}
\alias{hc_plot_hdi}
\title{Plot HDI}
\usage{
hc_plot_hdi(happy_hdi, title = NULL)
}
\arguments{
\item{happy_hdi}{A \code{data.frame} obtained with \code{estimate_hdi()}.}

\item{title}{Title. Default: \code{NULL}.}
}
\description{
Plot distribution of success rates and HDI estimates for
a given subset from a \code{data.frame}.
}
\examples{

\dontrun{
  e <- estimate_hdi(df = d, successes_col = successes_col, totals_col = totals_col, 
  group_cols = group_cols, aggregate_only = FALSE, sample_size = 1000)
  h <- e \%>\% dplyr::filter(Subset == "high.at")
  hc_plot_hdi(h, title = "PCR-Free vs Nano high.at")
}

}
