% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{hc_plot_roc}
\alias{hc_plot_roc}
\title{Plot a ROC curve for precision vs recall}
\usage{
hc_plot_roc(happy_roc, type, filter, subtype = "*", subset = "*")
}
\arguments{
\item{happy_roc}{A \code{happy_roc} object.}

\item{type}{Variant type.}

\item{filter}{Variant filter.}

\item{subtype}{Variant subtype. Default: \code{*}.}

\item{subset}{Variant subset. Default: \code{*}.}
}
\description{
Plot precision vs recall ROC curves from the selected filters. If \code{pr.all} is
provided, overlay pass points to \code{ALL} curve.
}
\examples{

\dontrun{
roc <- extract_metrics(happy_compare, table = "pr.all")
hc_plot_roc(roc, type = "INDEL", filter = "PASS")
}
}
