% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_roc.R
\name{plot.happy_roc}
\alias{plot.happy_roc}
\title{plot}
\usage{
\method{plot}{happy_roc}(x, filter = "PASS", type, subtype = "*", ...)
}
\arguments{
\item{x}{A `happy_roc` object.}

\item{filter}{Variant filter. Default = "PASS".}

\item{type}{Variant type. One of: "SNP", "INDEL".}

\item{subtype}{Variant subtype. Default = "*".}

\item{...}{Extra arguments.}
}
\description{
Plot ROC curves.
}

