% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{plot}
\alias{plot}
\alias{plot}
\title{Plot a ROC curve for precision vs recall}
\usage{
plot(happy_hdi, ...)

plot(happy_hdi, ...)
}
\arguments{
\item{happy_hdi}{A \code{data.frame} obtained with \code{estimate_hdi()}.}

\item{happy_roc}{A \code{happy_roc} object.}

\item{filter}{Variant filter. Default: \code{PASS}.}

\item{type}{Variant type. One of: \code{SNP}, \code{INDEL}.}

\item{subtype}{Variant subtype. Default: \code{*}.}

\item{subset}{Variant subset. Default: \code{*}.}

\item{legend_position}{A character string to pass to \code{ggplot} to specify
the legend position. Default: \code{bottom}.}

\item{title}{Title. Default: \code{NULL}.}
}
\value{
A \code{ggplot2} plot.

A \code{ggplot2} plot.
}
\description{
Plot precision vs recall ROC curves from the selected filters.

Plot distribution of success rates and HDI estimates for 
a given subset from a \code{data.frame}.
}
\examples{

\dontrun{
roc = extract(happy_compare, table = "pr_curve.all")
plot(roc, type = "SNP", filter = "ALL")  
}

\dontrun{
hdi = estimate_hdi(happy_extended, successes_col = "TRUTH.TP", totals_col = "TRUTH.TOTAL", 
                   group_cols = c("Group.Id", "Subset", "Type", "Subtype"))
plot(happy_hdi = hdi \%>\% filter(Subset == "*"))  
}
}
