% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happyCompare_list.R
\name{read_samplesheet_}
\alias{read_samplesheet_}
\title{Manually create happyCompare_list objects}
\usage{
read_samplesheet_(samplesheet, happy_results, ids)
}
\arguments{
\item{samplesheet}{A happyCompare samplesheet (\code{data.frame}). Required fields:
\code{Group.Id}, \code{Sample.Id}, \code{Replicate.Id},
\code{happy_prefix}.}

\item{happy_results}{A \code{happy_results_list} object obtained with \code{happyR}.}

\item{ids}{A \code{vector} of unique ids to map samplesheet entries with happy results.}
}
\value{
A \code{happyCompare_list} object, with the following fields: 
\itemize{ 
  \item{\code{samplesheet}: the original samplesheet, stored as a
        \code{data.frame}}. 
  \item{\code{happy_results}: a \code{happy_results_list}
        object that contains individual \code{happy_result} objects as defined in
        \code{happyR}}. 
}
}
\description{
Create a happyCompare_list object from its individual components.
}
\examples{

\dontrun{
happyCompare_list = read_samplesheet_(samplesheet = samplesheet, 
happy_results = happy_results, ids = ids)
}

}
