% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_summary.R
\name{summary.happy_summary}
\alias{summary.happy_summary}
\title{summary.happy_summary}
\usage{
\method{summary}{happy_summary}(object, type, filter = "PASS", digits = 4,
  kable_format = "markdown", aggregate = TRUE, colnames = NULL, ...)
}
\arguments{
\item{object}{A `happy_summary` object.}

\item{type}{Variant type. One of: SNP, INDEL.}

\item{filter}{Variant filter. One of: PASS (default), ALL.}

\item{digits}{Number of significant digits in summary statistics. Default: 4.}

\item{kable_format}{A character string to pass to `knitr::kable()`. Default: markdown.}

\item{aggregate}{Summarise performance across groups. Default: TRUE.}

\item{colnames}{Column names for output table. Default: happy_summary headers.}

\item{...}{Extra arguments.}
}
\description{
Summarise happy summary results into tabular format.
}

