% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/happy_extended.R
\name{tidy.happy_extended}
\alias{tidy.happy_extended}
\title{tidy}
\usage{
\method{tidy}{happy_extended}(x, ...)
}
\arguments{
\item{x}{A `happy_extended` object.}

\item{...}{Extra arguments.}
}
\value{
A `data.table` object.
}
\description{
Tidy a `happy_extended` object by converting into a single `data.table`.
}

